Moved dialogs presentation into window
Still need to clean it up and re-add
the accelerators.
