+++ data/gtk/window.ui
@@ -32,7 +32,7 @@
 		<section>
 			<item>
 				<attribute name="label" translatable="yes">_Preferences</attribute>
-				<attribute name="action">app.preferences</attribute>
+				<attribute name="action">win.preferences</attribute>
 			</item>
 			<item>
 				<attribute name="label" translatable="yes">_Keyboard Shortcuts</attribute>
@@ -40,7 +40,7 @@
 			</item>
 			<item>
 				<attribute name="label" translatable="yes">_About Add Water</attribute>
-				<attribute name="action">app.about</attribute>
+				<attribute name="action">win.about</attribute>
 			</item>
 		</section>
 	</menu>
